Add react-mixin to support using mixins with classes. This allows us to use some of timing functions. Also Refactored a bit more to move component code into separate files and just make the index files 'public-api'
