+++ package.json
@@ -8,6 +8,7 @@
   "dependencies": {
     "lodash": "^4.3.0",
     "ngeohash": "^0.6.0",
+    "react-mixin": "^2.0.2",
     "react-native": "^0.19.0",
     "react-native-asset-library-to-base64": "^1.0.1",
     "react-native-camera": "^0.3.8",
