+++ src/components/button/index.js
@@ -1,14 +1,3 @@
-import React, {
-  TouchableHighlight,
-  Text
-} from 'react-native';
-
-import styles from './styles';
-
-export default (props) => {
-  return (
-    <TouchableHighlight style={styles.baseContainer} onPress={props.onPress}>
-      <Text style={styles.baseType}>{props.children}</Text>
-    </TouchableHighlight>
-  );
-}
+'use strict';
+import Button from './button.component.js';
+export { Button };
