+++ src/components/button/styles.js
@@ -1,12 +1,12 @@
 import { StyleSheet } from 'react-native';
-import { brandColor, buttonTypeColor, h4TypeSize } from '../../styles/config';
+import { buttonBGColor, buttonTypeColor, h4TypeSize } from '../../styles/config';
 
 export default StyleSheet.create({
   baseContainer: {
     height: 40,
     justifyContent: 'center',
     alignItems: 'center',
-    backgroundColor: brandColor,
+    backgroundColor: buttonBGColor,
   },
   baseType: {
     color: buttonTypeColor,
