+++ src/components/h1/H1.component.js
@@ -1,0 +1,6 @@
+import React, { Text } from 'react-native';
+import styles from './styles';
+
+export default (props) => {
+  return <Text style={[styles.h1, props.style]}>{props.children}</Text>;
+};
