+++ src/components/h1/index.js
@@ -1,6 +1,3 @@
-import React, { Text } from 'react-native';
-import styles from './styles';
-
-export default (props) => {
-  return <Text style={[styles.h1, props.style]}>{props.children}</Text>;
-};
+'use strict';
+import H1 from './H1.component.js';
+export { H1 };
