+++ src/config/routes.js
@@ -1,5 +1,5 @@
 'use strict';
-import SplashPage from '../components/splash-page';
+import { SplashPage } from '../components/splash-page';
 import LoginPage from '../containers/login.container';
 import TrailMap from '../containers/trail-map.container';
 import ReportProblem from '../containers/report-problem.container';
