+++ src/containers/trail-map.container.js
@@ -1,7 +1,7 @@
 'use strict';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import TrailMap from '../components/trail-map';
+import { TrailMap } from '../components/trail-map';
 import { fetchAnnotations } from '../actions/trail-annotations.action-creator';
 import { fetchTrails } from '../actions/trails.action-creator';
